Move read_metric_file and fetch_participant_id functions to utils.py
